[BE] Update: FullcourseDetail 수정

See merge request s07-bigdata-dist-sub2/S07P22E106!233

+++ backend/src/main/java/com/ssafy/fullcourse/domain/place/controller/PlaceController.java
@@ -30,7 +30,7 @@ public class PlaceController {
     public ResponseEntity<BaseResponseBody> listPlace(@PathVariable String type, Pageable pageable,
                                                        @RequestParam(required = false, defaultValue = "") String keyword,
                                                        @RequestParam(required = false, defaultValue = "") String tag,
-                                                       @RequestParam(required = false, defaultValue = "0") Integer maxDist,
+                                                       @RequestParam(required = false, defaultValue = "0") Float maxDist,
                                                        @RequestParam(required = false, defaultValue = "0") Float recentLat,
                                                        @RequestParam(required = false, defaultValue = "0") Float recentLng
     ) throws Exception {
